Load systray icon from resource compiled into executable
This lets us get rid of having to create a temp directory to store the icon
as well as the weird dance of creating a temp file from data already stored
within the executable.

FIXME: As we don't create the temp dir anymore, this breaks icons for menu
items at the moment.